+++ qt/gen_qcolorspace.h
@@ -26,28 +26,28 @@ typedef struct QPointF QPointF;
 #endif
 
 QColorSpace* QColorSpace_new();
-QColorSpace* QColorSpace_new2(uintptr_t namedColorSpace);
-QColorSpace* QColorSpace_new3(uintptr_t primaries, uintptr_t transferFunction);
-QColorSpace* QColorSpace_new4(uintptr_t primaries, float gamma);
-QColorSpace* QColorSpace_new5(QPointF* whitePoint, QPointF* redPoint, QPointF* greenPoint, QPointF* bluePoint, uintptr_t transferFunction);
+QColorSpace* QColorSpace_new2(int namedColorSpace);
+QColorSpace* QColorSpace_new3(int primaries, int transferFunction);
+QColorSpace* QColorSpace_new4(int primaries, float gamma);
+QColorSpace* QColorSpace_new5(QPointF* whitePoint, QPointF* redPoint, QPointF* greenPoint, QPointF* bluePoint, int transferFunction);
 QColorSpace* QColorSpace_new6(QColorSpace* colorSpace);
-QColorSpace* QColorSpace_new7(uintptr_t primaries, uintptr_t transferFunction, float gamma);
-QColorSpace* QColorSpace_new8(QPointF* whitePoint, QPointF* redPoint, QPointF* greenPoint, QPointF* bluePoint, uintptr_t transferFunction, float gamma);
+QColorSpace* QColorSpace_new7(int primaries, int transferFunction, float gamma);
+QColorSpace* QColorSpace_new8(QPointF* whitePoint, QPointF* redPoint, QPointF* greenPoint, QPointF* bluePoint, int transferFunction, float gamma);
 void QColorSpace_OperatorAssign(QColorSpace* self, QColorSpace* colorSpace);
 void QColorSpace_Swap(QColorSpace* self, QColorSpace* colorSpace);
-uintptr_t QColorSpace_Primaries(const QColorSpace* self);
-uintptr_t QColorSpace_TransferFunction(const QColorSpace* self);
+int QColorSpace_Primaries(const QColorSpace* self);
+int QColorSpace_TransferFunction(const QColorSpace* self);
 float QColorSpace_Gamma(const QColorSpace* self);
-void QColorSpace_SetTransferFunction(QColorSpace* self, uintptr_t transferFunction);
-QColorSpace* QColorSpace_WithTransferFunction(const QColorSpace* self, uintptr_t transferFunction);
-void QColorSpace_SetPrimaries(QColorSpace* self, uintptr_t primariesId);
+void QColorSpace_SetTransferFunction(QColorSpace* self, int transferFunction);
+QColorSpace* QColorSpace_WithTransferFunction(const QColorSpace* self, int transferFunction);
+void QColorSpace_SetPrimaries(QColorSpace* self, int primariesId);
 void QColorSpace_SetPrimaries2(QColorSpace* self, QPointF* whitePoint, QPointF* redPoint, QPointF* greenPoint, QPointF* bluePoint);
 bool QColorSpace_IsValid(const QColorSpace* self);
 QColorSpace* QColorSpace_FromIccProfile(QByteArray* iccProfile);
 QByteArray* QColorSpace_IccProfile(const QColorSpace* self);
 QColorTransform* QColorSpace_TransformationToColorSpace(const QColorSpace* self, QColorSpace* colorspace);
-void QColorSpace_SetTransferFunction2(QColorSpace* self, uintptr_t transferFunction, float gamma);
-QColorSpace* QColorSpace_WithTransferFunction2(const QColorSpace* self, uintptr_t transferFunction, float gamma);
+void QColorSpace_SetTransferFunction2(QColorSpace* self, int transferFunction, float gamma);
+QColorSpace* QColorSpace_WithTransferFunction2(const QColorSpace* self, int transferFunction, float gamma);
 void QColorSpace_Delete(QColorSpace* self);
 
 #ifdef __cplusplus
