+++ qt/gen_qcryptographichash.h
@@ -23,14 +23,14 @@ typedef struct QCryptographicHash QCryptographicHash;
 typedef struct QIODevice QIODevice;
 #endif
 
-QCryptographicHash* QCryptographicHash_new(uintptr_t method);
+QCryptographicHash* QCryptographicHash_new(int method);
 void QCryptographicHash_Reset(QCryptographicHash* self);
 void QCryptographicHash_AddData(QCryptographicHash* self, const char* data, int length);
 void QCryptographicHash_AddDataWithData(QCryptographicHash* self, QByteArray* data);
 bool QCryptographicHash_AddDataWithDevice(QCryptographicHash* self, QIODevice* device);
 QByteArray* QCryptographicHash_Result(const QCryptographicHash* self);
-QByteArray* QCryptographicHash_Hash(QByteArray* data, uintptr_t method);
-int QCryptographicHash_HashLength(uintptr_t method);
+QByteArray* QCryptographicHash_Hash(QByteArray* data, int method);
+int QCryptographicHash_HashLength(int method);
 void QCryptographicHash_Delete(QCryptographicHash* self);
 
 #ifdef __cplusplus
