+++ qt/gen_qdeadlinetimer.h
@@ -20,17 +20,17 @@ typedef struct QDeadlineTimer QDeadlineTimer;
 #endif
 
 QDeadlineTimer* QDeadlineTimer_new();
-QDeadlineTimer* QDeadlineTimer_new2(uintptr_t param1);
+QDeadlineTimer* QDeadlineTimer_new2(int param1);
 QDeadlineTimer* QDeadlineTimer_new3(long long msecs);
 QDeadlineTimer* QDeadlineTimer_new4(QDeadlineTimer* param1);
-QDeadlineTimer* QDeadlineTimer_new5(uintptr_t type_);
-QDeadlineTimer* QDeadlineTimer_new6(uintptr_t param1, uintptr_t type_);
-QDeadlineTimer* QDeadlineTimer_new7(long long msecs, uintptr_t typeVal);
+QDeadlineTimer* QDeadlineTimer_new5(int type_);
+QDeadlineTimer* QDeadlineTimer_new6(int param1, int type_);
+QDeadlineTimer* QDeadlineTimer_new7(long long msecs, int typeVal);
 void QDeadlineTimer_Swap(QDeadlineTimer* self, QDeadlineTimer* other);
 bool QDeadlineTimer_IsForever(const QDeadlineTimer* self);
 bool QDeadlineTimer_HasExpired(const QDeadlineTimer* self);
-uintptr_t QDeadlineTimer_TimerType(const QDeadlineTimer* self);
-void QDeadlineTimer_SetTimerType(QDeadlineTimer* self, uintptr_t typeVal);
+int QDeadlineTimer_TimerType(const QDeadlineTimer* self);
+void QDeadlineTimer_SetTimerType(QDeadlineTimer* self, int typeVal);
 long long QDeadlineTimer_RemainingTime(const QDeadlineTimer* self);
 long long QDeadlineTimer_RemainingTimeNSecs(const QDeadlineTimer* self);
 void QDeadlineTimer_SetRemainingTime(QDeadlineTimer* self, long long msecs);
@@ -44,13 +44,13 @@ QDeadlineTimer* QDeadlineTimer_Current();
 QDeadlineTimer* QDeadlineTimer_OperatorPlusAssign(QDeadlineTimer* self, long long msecs);
 QDeadlineTimer* QDeadlineTimer_OperatorMinusAssign(QDeadlineTimer* self, long long msecs);
 void QDeadlineTimer_OperatorAssign(QDeadlineTimer* self, QDeadlineTimer* param1);
-void QDeadlineTimer_SetRemainingTime2(QDeadlineTimer* self, long long msecs, uintptr_t typeVal);
+void QDeadlineTimer_SetRemainingTime2(QDeadlineTimer* self, long long msecs, int typeVal);
 void QDeadlineTimer_SetPreciseRemainingTime2(QDeadlineTimer* self, long long secs, long long nsecs);
-void QDeadlineTimer_SetPreciseRemainingTime3(QDeadlineTimer* self, long long secs, long long nsecs, uintptr_t typeVal);
-void QDeadlineTimer_SetDeadline2(QDeadlineTimer* self, long long msecs, uintptr_t timerType);
+void QDeadlineTimer_SetPreciseRemainingTime3(QDeadlineTimer* self, long long secs, long long nsecs, int typeVal);
+void QDeadlineTimer_SetDeadline2(QDeadlineTimer* self, long long msecs, int timerType);
 void QDeadlineTimer_SetPreciseDeadline2(QDeadlineTimer* self, long long secs, long long nsecs);
-void QDeadlineTimer_SetPreciseDeadline3(QDeadlineTimer* self, long long secs, long long nsecs, uintptr_t typeVal);
-QDeadlineTimer* QDeadlineTimer_Current1(uintptr_t timerType);
+void QDeadlineTimer_SetPreciseDeadline3(QDeadlineTimer* self, long long secs, long long nsecs, int typeVal);
+QDeadlineTimer* QDeadlineTimer_Current1(int timerType);
 void QDeadlineTimer_Delete(QDeadlineTimer* self);
 
 #ifdef __cplusplus
